fix: send normal msg lark bot
